insertEnd and insertBegin print directly
For the CSV database type, it is important that there are no
extraneous empty lines. Unfortunately, the current interface
required an `insertEnd` and `insertBegin` that always printed
(at mininimum) a newline.

With this diff, `insertEnd` and `insertBegin` behave like
`comment`: they print directly for relevant database types and
are a noop for others.